pyproject.toml
[project]
name = "spark-llm-agent-shell"
version = "0.1.0"
description = "Policy-driven shell around an LLM for Spark optimization workflows."
requires-python = ">=3.10"
dependencies = [
  "PyYAML>=6.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.4",
  "pytest>=8.3.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
pythonpath = ["."]
